test: moved tests to new files
